+++ app/services/get_answers_for_steps.rb
@@ -22,12 +22,21 @@ class GetAnswersForSteps
       end
 
       if answer.respond_to?(:further_information)
-        hash["extended_answer_#{step.contentful_id}"] = [
-          {
-            "response" => answer.response,
-            "further_information" => answer.further_information
-          }
-        ]
+        hash["extended_answer_#{step.contentful_id}"] = if answer.further_information.is_a?(Hash)
+          answer.further_information.each_pair.each_with_object([]) do |(key, value), array|
+            array << {
+              "response" => key.gsub("_further_information", "").capitalize,
+              "further_information" => value
+            }
+          end
+        else
+          [
+            {
+              "response" => answer.response,
+              "further_information" => answer.further_information
+            }
+          ]
+        end
       end
 
       hash["answer_#{step.contentful_id}"] = answer.response.to_s