+++ CHANGELOG.md
@@ -33,6 +33,7 @@ The format is based on [Keep a Changelog 1.0.0].
 - create extended further information text box option for checkbox answers
 - radio questions can be configured with an "or" separator
 - specification can show further information for radio answers
+- specification can show further information for checkbox answers
 
 ## [release-005] - 2021-1-19